Adding DTO's to the service
