+++ mobi-order-processing/mobi-order-service/pom.xml
@@ -50,6 +50,12 @@
 			<version>1.18.10</version>
 			<scope>provided</scope>
 		</dependency>
+		<dependency>
+			<groupId>com.mobi.demo</groupId>
+			<artifactId>mobi-inventory-service</artifactId>
+			<version>0.0.1-SNAPSHOT</version>
+			<scope>compile</scope>
+		</dependency>
 	</dependencies>
 
 </project>
