+++ mobi-order-processing/mobi-order-service/src/main/java/com/mobi/order/mobiOrderService/dto/OrderDto.java
@@ -1,0 +1,14 @@
+package com.mobi.order.mobiOrderService.dto;
+
+
+import com.mobi.inventory.dto.ProductDto;
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class OrderDto {
+    long orderId;
+    long customerId;
+    List<ProductDto> products;
+}
