+++ mobi-order-processing/mobi-order-service/src/main/java/com/mobi/order/mobiOrderService/dto/ResponseDto.java
@@ -1,0 +1,7 @@
+package com.mobi.order.mobiOrderService.dto;
+
+public class ResponseDto {
+    String message;
+    String status;
+    OrderDto orderDto;
+}
