--- mobi-order-processing/mobi-order-service/src/main/java/com/mobi/order/mobiOrderService/models/RequestDto.java
@@ -1,11 +1,0 @@
-package com.mobi.order.mobiOrderService.models;
-
-import com.mobi.order.mobiOrderService.entities.Order;
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-public class RequestDto {
-    List<Order> orderList;
-}
