--- mobi-order-processing/mobi-order-service/src/main/java/com/mobi/order/mobiOrderService/models/ResponseDto.java
@@ -1,11 +1,0 @@
-package com.mobi.order.mobiOrderService.models;
-
-import com.mobi.order.mobiOrderService.entities.Order;
-
-import java.util.List;
-
-public class ResponseDto {
-    String message;
-    String status;
-    List<Order> orderList;
-}
