+++ mobi-order-processing/mobi-order-service/src/main/java/com/mobi/order/mobiOrderService/services/OrderService.java
@@ -1,15 +1,30 @@
-//package com.mobi.order.mobiOrderService.services;
-//
-//import com.mobi.order.mobiOrderService.entities.OrderDetails;
-//import com.mobi.order.mobiOrderService.models.PlaceOrderReq;
-//import com.mobi.order.mobiOrderService.models.Product;
-//
-//import java.util.List;
-//
-//public interface OrderService {
-//
-//    OrderDetails getOrder(Long orderId);
-//    Long placeOrder(PlaceOrderReq placeOrderReq);
-//    String getOrderStatus(Long orderId);
-//
-//}
+package com.mobi.order.mobiOrderService.services;
+
+import com.mobi.order.mobiOrderService.entities.Order;
+import com.mobi.order.mobiOrderService.dto.OrderDto;
+import com.mobi.order.mobiOrderService.repository.OrderRepository;
+//import com.mobi.order.mobiOrderService.services.OrderService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+@Service
+public class OrderService {
+
+    @Autowired
+    OrderRepository orderRepository;
+
+//    @Override
+    public Order getOrder(Long orderId) {
+        return orderRepository.findByOrderId(orderId);
+    }
+
+//    @Override
+    public Long placeOrder(OrderDto orderDto) {
+        return null;
+    }
+
+//    @Override
+    public String getOrderStatus(Long orderId) {
+        return null;
+    }
+}
